+++ core/translate/planner.rs
@@ -1,11 +1,11 @@
-use super::plan::{Aggregate, BTreeTableReference, DeletePlan, Plan, SourceOperator};
+use super::plan::{Aggregate, BTreeTableReference, SourceOperator};
 use crate::{
     function::Func,
     schema::Schema,
     util::{exprs_are_equivalent, normalize_ident},
     Result,
 };
-use sqlite3_parser::ast::{self, Expr, FromClause, JoinType, Limit, QualifiedName};
+use sqlite3_parser::ast::{self, Expr, FromClause, JoinType, Limit};
 
 pub struct OperatorIdCounter {
     id: usize,
@@ -251,49 +251,6 @@ pub fn bind_column_references(
         }
         ast::Expr::Variable(_) => todo!(),
     }
-}
-
-pub fn prepare_delete_plan(
-    schema: &Schema,
-    tbl_name: &QualifiedName,
-    where_clause: Option<Expr>,
-    limit: Option<Limit>,
-) -> Result<Plan> {
-    let table = match schema.get_table(tbl_name.name.0.as_str()) {
-        Some(table) => table,
-        None => crate::bail_corrupt_error!("Parse error: no such table: {}", tbl_name),
-    };
-
-    let table_ref = BTreeTableReference {
-        table: table.clone(),
-        table_identifier: table.name.clone(),
-        table_index: 0,
-    };
-    let referenced_tables = vec![table_ref.clone()];
-
-    // Parse the WHERE clause
-    let resolved_where_clauses = parse_where(where_clause, &[table_ref.clone()])?;
-
-    // Parse the LIMIT clause
-    let resolved_limit = limit.and_then(parse_limit);
-
-    let plan = DeletePlan {
-        source: SourceOperator::Scan {
-            id: 0,
-            table_reference: table_ref.clone(),
-            predicates: resolved_where_clauses.clone(),
-            iter_dir: None,
-        },
-        result_columns: vec![],
-        where_clause: resolved_where_clauses,
-        order_by: None,
-        limit: resolved_limit,
-        referenced_tables,
-        available_indexes: vec![],
-        contains_constant_false_condition: false,
-    };
-
-    Ok(Plan::Delete(plan))
 }
 
 pub fn parse_from(
